Set up your R Markdown document to use DataCamp light

Enable hooks and parsing utilities so that special code chunks where the ex and type options are specified, are converted into interactive exercises on the resulting HTML page. Use build_example to generate an example document. Simply knitting this document will generate a document with DataCamp Light powered code chunks.

go_interactive(greedy = TRUE, height = 300)

Arguments

greedy
whether or not to 'greedily' convert code chunks into DataCamp Light frames.

height
height in pixels that you want the resulting DataCamp Light frame to have.

Details

With the greedy argument, you can control which elements of your R Markdown document are converted into DataCamp Light chunks. By default greedy is TRUE, leading to all R code chunks to be converted to interactive frames. Only chunks for which you specify the option tut = FALSE are not converted. If `greedy` is FALSE, only chunks for which you specify tut = TRUE are converted.